% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scrm.R
\docType{package}
\name{scrm-package}
\alias{scrm-package}
\title{Simulating the evolution of biological sequences}
\description{
The Sequential Coalescent with Recombination Model (SCRM) is an
approximation of the Ancestral Recombination Graph. It can be used to
simulate the neutral evolution of chromosomes/biological sequences
subject to possibly complicated population structure.
The program \emph{scrm} is an implementation of this model that is designed to
act as an drop-in replacement for the widely adopted coalescent simulator
\emph{ms}. This package contains \emph{scrm} along with a convinient R interface.
}
\author{
Paul Staab,
Zhu Sha,
Dirk Metzler &
Gerton Lunter

Maintainer: Paul Staab \email{develop@paulstaab.de}
}
\seealso{
\itemize{
  \item \code{\link{scrm}} for details on how to use \emph{scrm},
  \item \code{vignette('scrm-Arguments')} for an overview of command line arguments and
  \item \code{vignette('scrm-TreesForApe')} for an example on using
        genealogies simulated with \emph{scrm} with package 'ape'.
}
}

