% Generated by roxygen2 (4.0.2): do not edit by hand
\name{scrm}
\alias{scrm}
\title{Simulation the evolution of biological sequences}
\usage{
scrm(args, file = "")
}
\arguments{
\item{args}{A string containing the command line arguments for scrm. Look at
scrms vignette for a description of available arguments.}

\item{file}{File provided, scrm will additionally write it is output into
the given file, using an ms-like text format.}
}
\description{
Simulation the evolution of biological sequences
}
\section{Seeding}{

The R version of scrm uses random number generated by R's random generator.
Therefore, its '-seed' argument will be ignored, and no seed is given in its
output. Use the R function \code{\link{set.seed}} prior to calling this
function to ensure reproduciblity of results.
}
\examples{
set.seed(17)
scrm('5 1 -r 10 100 -t 5')
\dontrun{scrm('5 1 -r 10 100 -t 5', file = tempfile('scrm_'))}
}

