% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{scrmr-package}
\alias{scrmr-package}
\title{Fast estimation of demographic parameters}
\description{
Fast estimation of demographic parameters
}
\details{
Jaatha is a composite maximum likelihood method to estimate parameters
of a speciation model of closely related biological species out of SNP
data.
}
\author{
Lisha Naduvilezhath \email{lisha (at) biologie.uni-muenchen.de},
Paul R. Staab \email{staab (at) biologie.uni-muenchen.de} and
Dirk Metzler \email{metzler (at) biologie.uni-muenchen.de}

Maintainer: Paul R. Staab \email{staab (at) biologie.uni-muenchen.de}
}
\keyword{package}

