% Generated by roxygen2 (4.0.2): do not edit by hand
\name{scrm}
\alias{scrm}
\title{Simulate the evolution of biological sequences}
\usage{
scrm(args, file = "")
}
\arguments{
\item{args}{A string containing the command line arguments for scrm. Look at
scrms vignette for a description of available arguments.}

\item{file}{If provided, scrm will additionally write it is output into
a file with the given file, using an ms-like text output.}
}
\value{
A named list of summary statistics.
}
\description{
Simulate the evolution of biological sequences
}
\section{Seeding}{

The R version of scrm uses random number generated by R's random generator.
Therefore, its '-seed' argument will be ignored, and no seed is given in its
output. Use the R function \code{\link{set.seed}} prior to calling this
function to ensure reproduciblity of results.
}
\examples{
set.seed(17)

# 5 Chromosomes with 100 bases each with recombination and mutation
sum_stats <- scrm('5 1 -r 5 100 -t 5 -T -L')
str(sum_stats)

# Simulate the site frequency spectrum at 3 loci. For each locus
# 10 Chromosomes of 1Mb length are sampled from two populations with
# migration inbetween.
scrm('10 3 -r 400 1000000 -l 100000 -I 2 4 6 0.5 -t 5 -oSFS')$sfs
\dontrun{scrm('5 1 -r 10 100 -t 5', file = tempfile('scrm_'))}
}

